disasm: fix pltgottargets handling when no matches are found
If you specify sep when splitting a string but the sep value isn't found,
you get a list with a single entry that is an empty string.
More generally, consecutive delimiters aren't grouped together and are
deemed to delimit empty strings.
An empty string is an invalid argument for the pltgotmapper function, so
filter them out.